+++ CHANGELOG.md
@@ -8,4 +8,10 @@ CHANGELOG
 
 1.1.0 (2013-09-16)
 ------------------
-* Added new method getRaw() in ConfigRepositoryInterface #1.
+* [New] Added new method getRaw() in ConfigRepositoryInterface #1.
+
+1.2.0
+-----
+* [New] *ConfigRepositoryOperationInterface*: Define the operations with repositories.
+* [New] Operations with repositories: *union* and *intersection* method was added to repository.
+* [Deprecated] Method *mergeWith* of repository is deprecated and replaced by *union* method.